requirements.txt
streamlit
pandas
altair
plotly
numpy
duckdb
openai
python-dotenv
